requirements-lite.txt
# Minimal dependencies for running:
#   - run/collect_miro_issues.sh  (Stage I: data collection)
#   - run/setup_miro_envs.sh     (Stage II: env setup)

# Stage I: data collection
beautifulsoup4>=4.12
fastcore>=1.5
ghapi>=1.0
Pygments>=2.17
requests>=2.31
tqdm>=4.66

# Stage II: env setup (multi-agent Dockerfile + eval script generation)
docker>=7.0
python-dotenv>=1.0
filelock>=3.13
httpx>=0.27
litellm>=1.44
loguru>=0.7
openai>=1.50
packaging>=23.2
tenacity>=8.2
